WIP on React concept

+++ src/components/utilities/fakedb.js
@@ -3,7 +3,12 @@
 const addToDb = (id) => {
   const exists = localStorage.getItem(id);
   console.log(exists);
-  localStorage.setItem(id, 1);
+  if (!exists) {
+    localStorage.setItem(id, 1);
+  } else {
+    const newCount = parseInt(exists) + 1;
+    localStorage.setItem(id, newCount);
+  }
 };
 
 export { addToDb };
